// src/types/generated/PresentationEditResponseMetadata.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PresentationSlideText } from "./PresentationSlideText";

export type PresentationEditResponseMetadata = { hasData: boolean, slideCount: number, currentSlideIndex: number, slide?: PresentationSlideText, converter: string, normalizedWmfCount: number, };